Move methods to wallet interface
