+++ src/main/java/eu/qwan/vender/Wallet.java
@@ -4,4 +4,6 @@ public interface Wallet {
 
     boolean hasValue(int amount);
     void reduce(int amount);
+    int getCredits();
+    void addCredits(int amount);
 }
