+++ src/test/java/eu/qwan/vender/VendingMachineTest.java
@@ -99,7 +99,7 @@ public class VendingMachineTest {
 	@Test
 	public void Testcheckout_chip_if_chipknip_inserted() {
 		machine.configure(Choice.sprite, Can.sprite, 1, 1);
-		Chipknip chip = new Chipknip(10);
+		ChipknipWallet chip = new ChipknipWallet(10);
 		machine.insertChip(chip);
 		assertEquals(Can.sprite, machine.deliver(Choice.sprite));
 		assertEquals(9, chip.credits);
@@ -108,7 +108,7 @@ public class VendingMachineTest {
 	@Test
 	public void Testcheckout_chip_empty() {
 		machine.configure(Choice.sprite, Can.sprite, 1, 1);
-		Chipknip chip = new Chipknip(0);
+		ChipknipWallet chip = new ChipknipWallet(0);
 		machine.insertChip(chip);
 		assertEquals(Can.none, machine.deliver(Choice.sprite));
 		assertEquals(0, chip.credits);
